+++ sdk/python/pulumi_aws/ec2transitgateway/__init__.py
@@ -11,6 +11,7 @@ from .route_table_propagation import *
 from .vpc_attachment import *
 from .vpc_attachment_accepter import *
 from .get_transit_gateway import *
+from .get_direct_connect_gateway_attachment import *
 from .get_route_table import *
 from .get_vpc_attachment import *
 from .get_vpn_attachment import *